tulip-python: Remove tulipgui packaging and deployment on PyPI
Distributing tulipgui-python on PyPI was a silly idea from my part.
Thinking it back, only tulip-python should be provided on PyPI.

The reasons of this removal are the following:

  - the module is not really used

  - bundling needed binaries in the wheel (notably the Qt ones) is painful
    and hard to maintain (especially on Linux)

  - the module is quite redundant with the main Tulip software, while being
    less intuituve to use

Previous users should now be redirected to use the main Tulip software for
visualization purposes or use the modules located in an official Tulip
installation instead.