Qt5: Migrate away from deprecated QtOpenGL module (WIP)
It is time to get rid of the use of the deprecated QtOpenGL module
to get better rendering performance through modern OpenGL use
but also ease future Tulip maintenance.

This is a breaking change so this needs deep testing before landing this to master.
Target Tulip release is 5.4 so this let some time for fixes and code improvements.