Remove awful and useless Tulip GUI tests
Software tests should be reproducible and easy to execute in order for contributors
to check their changes do not break anything. Those tests are clearly not as they
require to be run on Linux using a Gnome Shell desktop with HD screen resolution.
I managed to reproduce their execution environment but a lot of them was failing
and they took way too much time to execute.

Moreover their implementation is quite hackish and seems a nightmare to maintain.
Plus a lot of those tests could have been written using cppunit, like those
executing a graph algorithm and checking the results. What the point of implementing
a GUI test for that ? It would have been much better to write portable tests in
C++ instead.

Anyway, removing those tests once for all.