cmake: Remove the "without Python" build option
Finally remove that CMake option from the Tulip era. Python integration is a
great feature so why adding an option to remove it ? A quick look on the
SourceForge download stats of Tulip shows that installers with Python integration
are much more downloaded than the ones without it so always build Talipot with
Python from now on.