talipot-python: Allow Python IDE anchoring in the main Talipot window
Turning the Python IDE to a separate dialog was such a bad idea in terms of ergonomics.
So anchor it back in the right part of the main Talipot window by default.
The IDE can still be displayed in a separate dialog if needed through the use
of a new checkbox in its UI. The preference will be saved in the Talipot settings
to keep it at next application startup.