travis: Skip dmg bundle generation if job is about to timeout
The MacPorts travis job can take too much time to execute if the ccache
got invalidated. Thus, the job timeouts when generating the Talipot dmg bundle.
So skip that phase if the overall build took at least 45 minutes in order
for the job to exit with success and the updated ccache uploaded to speed
up next builds.