software: Remove Perspective plugin system from the Tulip era
Tulip was designed with a concept of Perspective plugins in order to implement
main window GUI for those who wanted to develop custom applications on top of
the Tulip framework. The main Tulip GUI was also implemented through that
system.

But that system is too much complex and not pleasant to use.
Developers who want to use Talipot Qt features should now handle creation of
a QApplication from their side and link to Talipot libraries.

So remove that no more needed complex plugin system and turn the previous
GraphPerspective plugin into a talipot executable. Also remove the
Talipot agent feature as it was not really useful and having a single
executable with the same name as the project is a better practice.