bundlers/win: Fix deployment and installer generation issues
As the old perspective system from Tulip has been removed, we now need
to ensure that the Python dll can be found when launching Talipot
(especially when Python is not globally installed to all users).
The solution was to create a simple executable retrieving the Python home
path dynamically and printing it to stdout.
Talipot is now run through a bat script adding the Python home path to the
PATH environment variable before starting the talipot executable

Fix again the Dll Hell issue with OpenSSL, Python and Qt, ensure Qt OpenSSL
is loaded prior to the Python one.

Ensure the required Qt5Svg dll for svg imageformats plugin is installed.

Ensure the NSIS installer can be generated when using MSYS2/MinGW or MSVC.

Cleanup the win_bundle.bat script.