talipot-ogl: Remove dependency to libjpeg and libpng
talipot-ogl was linked against libjpeg and libpng libraries for loading
image data to create OpenGL textures.

But that code was not really used anymore as the image loading are delegated
to Qt when using the Talipot GUI.

So remove the not really needed libjpeg and libpng dependency and replace their
use by the lightweight and public domain library stb_image which greatly does
the same job.