talipot-ogl/OpenGlConfigManager: Robustify OpenGL extension detection
Add a call to glewGetExtension to double check if an OpenGL extension is supported.

For instance on Linux using an Intel GPU, the GL_EXT_geometry_shader4 extension was
detected as supported by glewIsSupported but in fact it is not at all.
Hopefully glewGetExtension returns false in that case so better to check
twice than once.